Split the transport config away from the rule table
No rule reads [general] or [tls] — the proxy does, and only the proxy. Core's
Config narrows to the [rules.*] table it actually serves, which makes the
crate's reuse claim true: linting a capture no longer requires inventing a
listen address. The transport surface — GeneralConfig, TlsConfig, and the
h3-requires-TLS invariant — moves to the crate that enforces it, layered over
the rule table with serde(flatten) so one TOML file still feeds both layers,
and Deref so rule-layer call sites cross the boundary without spelling it.

Sections a layer does not own are ignored where they used to be required:
the narrowed config parses a full proxy file, and the proxy config asserts
both layers agree over the shipped example.
